+++ src/mol-gl/renderable/spheres.ts
@@ -24,6 +24,7 @@ export const SpheresSchema = {
     dXrayShaded: DefineSpec('string', ['off', 'on', 'inverted']),
     dTransparentBackfaces: DefineSpec('string', ['off', 'on', 'opaque']),
     dSolidInterior: DefineSpec('boolean'),
+    dClipPrimitive: DefineSpec('boolean'),
     dApproximate: DefineSpec('boolean'),
     uAlphaThickness: UniformSpec('f'),
     uBumpFrequency: UniformSpec('f', 'material'),