+++ src/mol-gl/shader/spheres.vert.ts
@@ -124,6 +124,13 @@ void main(void){
         }
     }
 
-    #include clip_instance
+    #if defined(dClipPrimitive) && !defined(dClipVariant_instance) && dClipObjectCount != 0
+        if (clipTest(vec4(vModelPosition.xyz, 0.0))) {
+            // move out of [ -w, +w ] to 'discard' in vert shader
+            gl_Position.z = 2.0 * gl_Position.w;
+        }
+    #else
+        #include clip_instance
+    #endif
 }
 `;